+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.1.0
+
+* rename codec class add Codec sufix
+* add static class Charset
+
 ## 0.0.1
 
 * include gbk,euc-kr,euc-jp,shift-jis,cp437,utf16/32  encode & decode